requirements.txt
transformers
torch
Pillow
groq
streamlit
fastapi
uvicorn
scikit-learn
python-dotenv
